+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+### Version 1.0.1
+
+- Now can start/stop auto disenchant from minimap button.
+- Now can open disenchant window when there are no items to disenchant (to clean ignored/session lists).
+- Added button to start/stop auto disenchant in the main disenchant window.
+- Bug fixes
+
 ### Version 1.0.0
 
-- CurseForge first release.
+- First release.
 
 ### Version 0.0.1